requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
langsmith
faiss-cpu
sentence-transformers
transformers
PyMuPDF